before add grad
